Refactored components and finished auth improvements
