+++ src/App.js
@@ -1,15 +1,12 @@
 import React from 'react';
 import { Route } from 'react-router-dom';
-import LoginPage from './components/LoginPage';
-import HomePage from './components/HomePage';
-import PassportPage from './components/PassportPage';
-import Dashboard from './components/Dashboard';
 import NavBar from './components/Navbar';
+import HomePage from './components/HomePage';
+import LoginPage from './components/signup/LoginPage';
+import PassportPage from './components/signup/PassportPage';
+import Dashboard from './components/dashboard/Dashboard';
 import FlashMessagesList from './components/flash/FlashMessagesList';
-import requireAuth from './utils/requireAuth';
 import './styles/App.css';
-
-import Profile from './components/Profile';
 
 class App extends React.Component {
   render() {
