+++ src/components/dashboard/Account.js
@@ -3,8 +3,10 @@ import React from 'react';
 export default class Account extends React.Component {
   render() {
     return (
-      <div>
-        <h1>This is the Account component.</h1>
+      <div className="community container">
+        <h1>
+          This is the Account Component
+        </h1>
       </div>
     );
   }
