+++ src/components/dashboard/Community.js
@@ -3,8 +3,10 @@ import React from 'react';
 export default class Community extends React.Component {
   render() {
     return (
-      <div>
-        <h1>This is the Community component.</h1>
+      <div className="community container">
+        <h1>
+          This is the Community Component
+        </h1>
       </div>
     );
   }
