+++ src/styles/App.scss
@@ -17,7 +17,7 @@
 // fcc FA icon in navbar
 .menu .fa-free-code-camp {
   margin-right: 8px;
-} 
+}
 
 .flash-message {
   margin: 20px 0;
@@ -37,4 +37,10 @@
 
 #dashboard-page-main-container {
   margin-bottom: 50px;
+}
+
+
+// styles for components:
+.community, .account {
+  text-align: center;
 }
